feat: seperate BlockInfo to independent file
